Implemented Requirements Update for Shuttle
Implemented a check to validate that the user is 5 minutes away from either concordia campus before being allowed to use the shuttle service. The shuttle service is automatically charted with the further campus as the destination, and the user's current location as the origin. For testing purposes, 3 hardcoded test starting points have been added to validate the function of all 3 edge cases, simply uncomment which one you want to use as the testing start point for that refresh of the application.